requirements.txt
# Run in cmdline: $ pip install -r requirements.txt
MouseInfo==0.1.3
Pillow==9.3.0
PyAutoGUI==0.9.53
keyboard==0.13.5
numpy==1.23.4
opencv-python==4.6.0.66
pywin32==305
